Add files via upload
All methods for the API appear to be funtioning correctly. There is one test for each method in the file 'FractionChecker.java'. The main tasks left are to add proper testing, and to check and add to the documentation for each function, so that my (presumably massive) userbase can use them easily.